🐛 fix(home): polish action tiles
