+++ android/Reguerta/app/src/main/res/values/strings.xml
@@ -92,9 +92,10 @@
     <string name="home_dashboard_order_state_not_started">Not started</string>
     <string name="home_dashboard_order_state_unconfirmed">Unconfirmed</string>
     <string name="home_dashboard_order_state_completed">Done</string>
-    <string name="home_dashboard_my_order_subtitle_edit">Edit this week</string>
-    <string name="home_dashboard_my_order_subtitle_review">Review and confirm</string>
-    <string name="home_dashboard_my_order_subtitle_completed">Confirmed order</string>
+    <string name="home_dashboard_my_order_subtitle_edit">Make order</string>
+    <string name="home_dashboard_my_order_subtitle_review">Confirm order</string>
+    <string name="home_dashboard_my_order_subtitle_completed">View/edit order</string>
+    <string name="home_dashboard_my_order_subtitle_last_order">View last order</string>
     <string name="home_dashboard_received_orders_subtitle">Received</string>
     <string name="home_shell_version_android_format">Android %1$s</string>
     <string name="home_shell_dev_marker">DEV</string>
