+++ ios/Reguerta/.swiftlint
@@ -10,28 +10,18 @@ excluded:
   - DerivedData
   - Pods
 
-disabled_rules:
-  - trailing_comma
 
 line_length:
   warning: 160
   error: 220
 
 file_length:
-  warning: 1000
-  error: 4000
+  warning: 600
+  error: 1200
 
 type_body_length:
   warning: 350
-  error: 2000
-
-function_body_length:
-  warning: 50
-  error: 300
-
-cyclomatic_complexity:
-  warning: 10
-  error: 100
+  error: 600
 
 type_name:
   min_length:
@@ -39,14 +29,14 @@ type_name:
     error: 2
   max_length:
     warning: 48
-    error: 80
+    error: 60
 
 identifier_name:
   min_length:
     warning: 2
     error: 1
   max_length:
-    warning: 60
-    error: 100
+    warning: 56
+    error: 64
 
 reporter: "xcode"
