+++ ios/Reguerta/Reguerta/Presentation/Access/Core/ReguertaImagePickerField.swift
@@ -189,7 +189,7 @@ private struct ReguertaCameraCaptureView: UIViewControllerRepresentable {
 
         func imagePickerController(
             _ picker: UIImagePickerController,
-            didFinishPickingMediaWithInfo info: [UIImagePickerController.InfoKey : Any]
+            didFinishPickingMediaWithInfo info: [UIImagePickerController.InfoKey: Any]
         ) {
             let image = info[.originalImage] as? UIImage
             onCapture(image)
