+++ ios/Reguerta/Reguerta/Presentation/Access/Home/ContentView+ReceivedOrdersRoute.swift
@@ -1,4 +1,3 @@
-import FirebaseFirestore
 import SwiftUI
 
 private enum ReceivedOrdersTab: String, CaseIterable, Identifiable {
@@ -17,32 +16,6 @@ private enum ReceivedOrdersTab: String, CaseIterable, Identifiable {
     }
 }
 
-enum ProducerOrderStatus: String, CaseIterable {
-    case unread
-    case read
-    case prepared
-    case delivered
-
-    static func from(_ rawValue: String?) -> ProducerOrderStatus {
-        guard let rawValue else { return .unread }
-        return ProducerOrderStatus(rawValue: rawValue.trimmingCharacters(in: .whitespacesAndNewlines).lowercased()) ?? .unread
-    }
-
-    var title: String {
-        switch self {
-        case .unread: return "Pendiente"
-        case .read: return "Pendiente"
-        case .prepared: return "Preparado"
-        case .delivered: return "Entregado"
-        }
-    }
-}
-
-struct ProducerStatusVisualStyle {
-    let container: Color
-    let border: Color
-}
-
 private enum ReceivedOrdersLoadState {
     case idle
     case loading
@@ -54,72 +27,6 @@ private enum ReceivedOrdersLoadState {
 private struct ReceivedOrdersWindow {
     let isEnabled: Bool
     let targetWeekKey: String
-}
-
-private struct ReceivedOrderLineRecord: Identifiable {
-    let id: String
-    let orderId: String
-    let consumerId: String
-    let consumerDisplayName: String
-    let productId: String
-    let productName: String
-    let productImageUrl: String?
-    let companyName: String
-    let packagingLine: String
-    let quantity: Double
-    let quantityUnitSingular: String
-    let quantityUnitPlural: String
-    let subtotal: Double
-
-    var dedupKey: String {
-        "\(orderId)|\(consumerId)|\(productId)"
-    }
-}
-
-private struct ReceivedOrdersProductRow: Identifiable {
-    let productId: String
-    let productName: String
-    let productImageUrl: String?
-    let companyName: String
-    let packagingLine: String
-    let totalQuantity: Double
-    let quantityUnitSingular: String
-    let quantityUnitPlural: String
-
-    var id: String { productId }
-
-    func quantityUnitLabel() -> String {
-        receivedOrdersIsApproximatelyOne(totalQuantity) ? quantityUnitSingular : quantityUnitPlural
-    }
-}
-
-private struct ReceivedOrdersMemberLine: Identifiable {
-    let id: String
-    let productName: String
-    let packagingLine: String
-    let quantity: Double
-    let quantityUnitSingular: String
-    let quantityUnitPlural: String
-    let subtotal: Double
-
-    func quantityUnitLabel() -> String {
-        receivedOrdersIsApproximatelyOne(quantity) ? quantityUnitSingular : quantityUnitPlural
-    }
-}
-
-private struct ReceivedOrdersMemberGroup: Identifiable {
-    let id: String
-    let orderId: String
-    let consumerDisplayName: String
-    let producerStatus: ProducerOrderStatus
-    let lines: [ReceivedOrdersMemberLine]
-    let total: Double
-}
-
-private struct ReceivedOrdersSnapshot {
-    let byProductRows: [ReceivedOrdersProductRow]
-    let byMemberGroups: [ReceivedOrdersMemberGroup]
-    let generalTotal: Double
 }
 
 // swiftlint:disable:next type_body_length
@@ -572,441 +479,4 @@ private func resolveReceivedOrdersWindow(
         isEnabled: consultaWindow.isConsultaPhase,
         targetWeekKey: consultaWindow.isConsultaPhase ? consultaWindow.previousWeekKey : currentWeekKey
     )
-}
-
-private func fetchReceivedOrdersSnapshotForProducer(
-    producerId: String,
-    targetWeekKey: String,
-    db: Firestore = Firestore.firestore(),
-    environment: ReguertaFirestoreEnvironment = ReguertaRuntimeEnvironment.currentFirestoreEnvironment
-) async throws -> ReceivedOrdersSnapshot? {
-    let firestorePath = ReguertaFirestorePath(environment: environment)
-    let readTargets = receivedOrderlineReadTargets(
-        firestorePath: firestorePath,
-        environment: environment
-    )
-    let lines = try await fetchReceivedOrderLines(
-        producerId: producerId,
-        targetWeekKey: targetWeekKey,
-        readTargets: readTargets,
-        db: db
-    )
-    guard !lines.isEmpty else { return nil }
-    let statusesByOrderId = try await fetchReceivedOrderStatusesByOrderId(
-        orderIds: lines.map(\.orderId),
-        producerId: producerId,
-        db: db,
-        environment: environment
-    )
-    let synchronizedStatuses = await synchronizeUnreadReceivedOrderStatuses(
-        statusesByOrderId: statusesByOrderId,
-        producerId: producerId,
-        db: db,
-        environment: environment
-    )
-
-    return buildReceivedOrdersSnapshot(from: lines, statusesByOrderId: synchronizedStatuses)
-}
-
-private func receivedOrderlineReadTargets(
-    firestorePath: ReguertaFirestorePath,
-    environment: ReguertaFirestoreEnvironment
-) -> [String] {
-    Array(Set([
-        firestorePath.collectionPath(.orderlines),
-        "\(environment.rawValue)/collections/orderLines",
-        "\(environment.rawValue)/collections/orderlines"
-    ]))
-}
-
-private func fetchReceivedOrderLines(
-    producerId: String,
-    targetWeekKey: String,
-    readTargets: [String],
-    db: Firestore
-) async throws -> [ReceivedOrderLineRecord] {
-    var dedupedLinesByKey: [String: ReceivedOrderLineRecord] = [:]
-    var hasSuccessfulRead = false
-    var lastError: Error?
-
-    for orderlinesPath in readTargets {
-        do {
-            let snapshot = try await db.collection(orderlinesPath)
-                .whereField("vendorId", isEqualTo: producerId)
-                .whereField("weekKey", isEqualTo: targetWeekKey)
-                .getDocuments()
-            hasSuccessfulRead = true
-            for document in snapshot.documents {
-                if let line = receivedOrderLineRecord(
-                    from: document.data(),
-                    fallbackDocumentID: document.documentID
-                ) {
-                    dedupedLinesByKey[line.dedupKey] = line
-                }
-            }
-        } catch {
-            lastError = error
-        }
-    }
-
-    if !hasSuccessfulRead, let lastError {
-        throw lastError
-    }
-
-    return dedupedLinesByKey.values.sorted { lhs, rhs in
-        if lhs.consumerDisplayName != rhs.consumerDisplayName {
-            return lhs.consumerDisplayName.localizedCaseInsensitiveCompare(rhs.consumerDisplayName) == .orderedAscending
-        }
-        return lhs.productName.localizedCaseInsensitiveCompare(rhs.productName) == .orderedAscending
-    }
-}
-
-private func synchronizeUnreadReceivedOrderStatuses(
-    statusesByOrderId: [String: ProducerOrderStatus],
-    producerId: String,
-    db: Firestore,
-    environment: ReguertaFirestoreEnvironment
-) async -> [String: ProducerOrderStatus] {
-    let unreadOrderIds = statusesByOrderId
-        .filter { $0.value == .unread }
-        .map(\.key)
-    guard !unreadOrderIds.isEmpty else {
-        return statusesByOrderId
-    }
-
-    let markedAsRead = await markReceivedOrdersAsRead(
-        orderIds: unreadOrderIds,
-        producerId: producerId,
-        db: db,
-        environment: environment
-    )
-    guard !markedAsRead.isEmpty else {
-        return statusesByOrderId
-    }
-
-    var synchronizedStatuses = statusesByOrderId
-    for orderId in markedAsRead {
-        synchronizedStatuses[orderId] = .read
-    }
-    return synchronizedStatuses
-}
-
-private func buildReceivedOrdersProductRows(
-    from lines: [ReceivedOrderLineRecord]
-) -> [ReceivedOrdersProductRow] {
-    Dictionary(grouping: lines, by: \.productId)
-        .compactMap { productId, grouped -> ReceivedOrdersProductRow? in
-            guard let first = grouped.first else { return nil }
-            let totalQuantity = grouped.reduce(0) { partial, line in partial + line.quantity }
-            return ReceivedOrdersProductRow(
-                productId: productId,
-                productName: first.productName,
-                productImageUrl: first.productImageUrl,
-                companyName: first.companyName,
-                packagingLine: first.packagingLine,
-                totalQuantity: totalQuantity,
-                quantityUnitSingular: first.quantityUnitSingular,
-                quantityUnitPlural: first.quantityUnitPlural
-            )
-        }
-        .sorted { lhs, rhs in
-            lhs.productName.localizedCaseInsensitiveCompare(rhs.productName) == .orderedAscending
-        }
-}
-
-private func buildReceivedOrdersMemberGroups(
-    from lines: [ReceivedOrderLineRecord],
-    statusesByOrderId: [String: ProducerOrderStatus]
-) -> [ReceivedOrdersMemberGroup] {
-    Dictionary(grouping: lines, by: { line in
-        "\(line.consumerId)|\(line.consumerDisplayName)"
-    })
-    .compactMap { key, grouped -> ReceivedOrdersMemberGroup? in
-        guard let first = grouped.first else { return nil }
-        let memberLines = grouped.map { line in
-            ReceivedOrdersMemberLine(
-                id: "\(line.orderId)|\(line.productId)",
-                productName: line.productName,
-                packagingLine: line.packagingLine,
-                quantity: line.quantity,
-                quantityUnitSingular: line.quantityUnitSingular,
-                quantityUnitPlural: line.quantityUnitPlural,
-                subtotal: line.subtotal
-            )
-        }.sorted { lhs, rhs in
-            lhs.productName.localizedCaseInsensitiveCompare(rhs.productName) == .orderedAscending
-        }
-        let total = memberLines.reduce(0) { partial, line in partial + line.subtotal }
-        return ReceivedOrdersMemberGroup(
-            id: key,
-            orderId: first.orderId,
-            consumerDisplayName: first.consumerDisplayName,
-            producerStatus: statusesByOrderId[first.orderId] ?? .unread,
-            lines: memberLines,
-            total: total
-        )
-    }
-    .sorted { lhs, rhs in
-        lhs.consumerDisplayName.localizedCaseInsensitiveCompare(rhs.consumerDisplayName) == .orderedAscending
-    }
-}
-
-private func receivedOrdersGeneralTotal(from lines: [ReceivedOrderLineRecord]) -> Double {
-    lines.reduce(0) { partial, line in
-        partial + line.subtotal
-    }
-}
-
-private func fetchReceivedOrderStatusesByOrderId(
-    orderIds: [String],
-    producerId: String,
-    db: Firestore = Firestore.firestore(),
-    environment: ReguertaFirestoreEnvironment = ReguertaRuntimeEnvironment.currentFirestoreEnvironment
-) async throws -> [String: ProducerOrderStatus] {
-    let dedupedOrderIds = Array(Set(orderIds)).filter(\.isNotEmpty)
-    guard !dedupedOrderIds.isEmpty else { return [:] }
-
-    let firestorePath = ReguertaFirestorePath(environment: environment)
-    let readTargets = Array(Set([
-        firestorePath.collectionPath(.orders),
-        "\(environment.rawValue)/collections/orders"
-    ]))
-    var statusesByOrderId: [String: ProducerOrderStatus] = [:]
-    var hasSuccessfulRead = false
-    var lastError: Error?
-
-    for ordersPath in readTargets {
-        for chunk in dedupedOrderIds.chunked(into: 10) {
-            do {
-                let snapshot = try await db.collection(ordersPath)
-                    .whereField(FieldPath.documentID(), in: chunk)
-                    .getDocuments()
-                hasSuccessfulRead = true
-                for document in snapshot.documents {
-                    statusesByOrderId[document.documentID] = receivedOrderStatus(
-                        from: document.data(),
-                        producerId: producerId
-                    )
-                }
-            } catch {
-                lastError = error
-            }
-        }
-    }
-
-    if !hasSuccessfulRead, let lastError {
-        throw lastError
-    }
-    return statusesByOrderId
-}
-
-private func updateReceivedOrderProducerStatus(
-    orderId: String,
-    producerId: String,
-    status: ProducerOrderStatus,
-    db: Firestore = Firestore.firestore(),
-    environment: ReguertaFirestoreEnvironment = ReguertaRuntimeEnvironment.currentFirestoreEnvironment,
-    nowMillis: Int64 = Int64(Date().timeIntervalSince1970 * 1_000)
-) async -> ReceivedOrderStatusWriteResult {
-    let firestorePath = ReguertaFirestorePath(environment: environment)
-    let writeTargets = Array(Set([
-        firestorePath.collectionPath(.orders),
-        "\(environment.rawValue)/collections/orders"
-    ]))
-    let nowTimestamp = Timestamp(date: Date(timeIntervalSince1970: TimeInterval(nowMillis) / 1_000))
-    var lastFailure: ReceivedOrderStatusWriteResult = .failure
-
-    for ordersPath in writeTargets {
-        do {
-            let orderRef = db.document("\(ordersPath)/\(orderId)")
-            try await orderRef.updateData([
-                "producerStatus": status.rawValue,
-                "producerStatusesByVendor.\(producerId)": status.rawValue,
-                "producerStatusUpdatedBy": producerId,
-                "updatedAt": nowTimestamp
-            ])
-            return .success
-        } catch {
-            lastFailure = receivedOrderStatusWriteResult(from: error)
-        }
-    }
-
-    return lastFailure
-}
-
-private func markReceivedOrdersAsRead(
-    orderIds: [String],
-    producerId: String,
-    db: Firestore = Firestore.firestore(),
-    environment: ReguertaFirestoreEnvironment = ReguertaRuntimeEnvironment.currentFirestoreEnvironment
-) async -> Set<String> {
-    var updatedOrderIds = Set<String>()
-    for orderId in Array(Set(orderIds)).filter(\.isNotEmpty) {
-        let updateResult = await updateReceivedOrderProducerStatus(
-            orderId: orderId,
-            producerId: producerId,
-            status: .read,
-            db: db,
-            environment: environment
-        )
-        if updateResult == .success {
-            updatedOrderIds.insert(orderId)
-        }
-    }
-    return updatedOrderIds
-}
-
-private func buildReceivedOrdersSnapshot(
-    from lines: [ReceivedOrderLineRecord],
-    statusesByOrderId: [String: ProducerOrderStatus]
-) -> ReceivedOrdersSnapshot {
-    let byProductRows = buildReceivedOrdersProductRows(from: lines)
-    let byMemberGroups = buildReceivedOrdersMemberGroups(
-        from: lines,
-        statusesByOrderId: statusesByOrderId
-    )
-
-    return ReceivedOrdersSnapshot(
-        byProductRows: byProductRows,
-        byMemberGroups: byMemberGroups,
-        generalTotal: receivedOrdersGeneralTotal(from: lines)
-    )
-}
-
-private func receivedOrderStatus(
-    from data: [String: Any],
-    producerId: String
-) -> ProducerOrderStatus {
-    if let statusesByVendor = data["producerStatusesByVendor"] as? [String: Any],
-       let statusValue = statusesByVendor[producerId] as? String {
-        return ProducerOrderStatus.from(statusValue)
-    }
-    return ProducerOrderStatus.from(data["producerStatus"] as? String)
-}
-
-private extension ReceivedOrdersSnapshot {
-    func withProducerStatus(orderId: String, status: ProducerOrderStatus) -> ReceivedOrdersSnapshot {
-        ReceivedOrdersSnapshot(
-            byProductRows: byProductRows,
-            byMemberGroups: byMemberGroups.map { group in
-                if group.orderId == orderId {
-                    return ReceivedOrdersMemberGroup(
-                        id: group.id,
-                        orderId: group.orderId,
-                        consumerDisplayName: group.consumerDisplayName,
-                        producerStatus: status,
-                        lines: group.lines,
-                        total: group.total
-                    )
-                }
-                return group
-            },
-            generalTotal: generalTotal
-        )
-    }
-}
-
-extension ProducerOrderStatus {
-    var visualStyle: ProducerStatusVisualStyle {
-        switch self {
-        case .unread:
-            return ProducerStatusVisualStyle(
-                container: Color(.systemGray6).opacity(0.82),
-                border: Color(.systemGray4)
-            )
-        case .read:
-            return ProducerStatusVisualStyle(
-                container: Color(.systemGray6).opacity(0.82),
-                border: Color(.systemGray4)
-            )
-        case .prepared:
-            return ProducerStatusVisualStyle(
-                container: Color(red: 1.0, green: 0.95, blue: 0.84),
-                border: Color(red: 0.84, green: 0.66, blue: 0.31)
-            )
-        case .delivered:
-            return ProducerStatusVisualStyle(
-                container: Color(red: 0.90, green: 0.97, blue: 0.90),
-                border: Color(red: 0.46, green: 0.64, blue: 0.44)
-            )
-        }
-    }
-}
-
-private func receivedOrderLineRecord(
-    from data: [String: Any],
-    fallbackDocumentID: String
-) -> ReceivedOrderLineRecord? {
-    let orderId = receivedOrderString(from: data["orderId"]) ?? fallbackDocumentID
-    let consumerId = receivedOrderString(from: data["userId"]) ?? "__consumer_unknown__"
-    let consumerDisplayName = receivedOrderString(from: data["consumerDisplayName"]) ?? consumerId
-    let productId = receivedOrderString(from: data["productId"]) ?? fallbackDocumentID
-    let productName = receivedOrderString(from: data["productName"]) ?? "Producto"
-    let companyName = receivedOrderString(from: data["companyName"]) ?? "Productor"
-    let productImageUrl = receivedOrderString(from: data["productImageUrl"])
-    let quantity = receivedOrderDouble(from: data["quantity"]) ?? 0
-    guard quantity > 0 else { return nil }
-    let subtotal = receivedOrderDouble(from: data["subtotal"])
-        ?? quantity * (receivedOrderDouble(from: data["priceAtOrder"]) ?? 0)
-    let quantityUnitSingular = receivedOrderString(from: data["packContainerName"])
-        ?? receivedOrderString(from: data["unitName"])
-        ?? "ud."
-    let quantityUnitPlural = receivedOrderString(from: data["packContainerPlural"])
-        ?? receivedOrderString(from: data["unitPlural"])
-        ?? quantityUnitSingular
-
-    return ReceivedOrderLineRecord(
-        id: "\(orderId)_\(productId)_\(consumerId)",
-        orderId: orderId,
-        consumerId: consumerId,
-        consumerDisplayName: consumerDisplayName,
-        productId: productId,
-        productName: productName,
-        productImageUrl: productImageUrl,
-        companyName: companyName,
-        packagingLine: receivedOrderPackagingLine(from: data),
-        quantity: quantity,
-        quantityUnitSingular: quantityUnitSingular,
-        quantityUnitPlural: quantityUnitPlural,
-        subtotal: subtotal
-    )
-}
-
-private func receivedOrderPackagingLine(from data: [String: Any]) -> String {
-    let containerName = receivedOrderString(from: data["packContainerName"])
-        ?? receivedOrderString(from: data["unitName"])
-        ?? ""
-    let quantity = (receivedOrderDouble(from: data["packContainerQty"])
-        ?? receivedOrderDouble(from: data["unitQty"])
-        ?? 1).myOrderUiDecimal
-    let unitLabel = receivedOrderString(from: data["packContainerAbbreviation"])
-        ?? receivedOrderString(from: data["packContainerPlural"])
-        ?? receivedOrderString(from: data["unitAbbreviation"])
-        ?? receivedOrderString(from: data["unitPlural"])
-        ?? receivedOrderString(from: data["unitName"])
-        ?? ""
-
-    return [containerName, quantity, unitLabel]
-        .filter(\.isNotEmpty)
-        .joined(separator: " ")
-}
-
-private func receivedOrderString(from value: Any?) -> String? {
-    guard let raw = value as? String else { return nil }
-    let trimmed = raw.trimmingCharacters(in: .whitespacesAndNewlines)
-    return trimmed.isEmpty ? nil : trimmed
-}
-
-private func receivedOrderDouble(from value: Any?) -> Double? {
-    if let number = value as? NSNumber {
-        return number.doubleValue
-    }
-    if let raw = value as? String {
-        return Double(raw.replacingOccurrences(of: ",", with: "."))
-    }
-    return nil
-}
-
-private func receivedOrdersIsApproximatelyOne(_ value: Double) -> Bool {
-    abs(value - 1) < 0.000_1
 }
